backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
chromadb
google-generativeai
requests
python-dotenv
pydantic
numpy
